First alpha release of the Tremor Mouse
The attenuation of the mouse movement is determined by the speed of the mouse.  There are three levels of attenuation.  High speed with no attenuation.  Medium speed with a 2:1 attenuation and Low speed with a 4:1  attenuation.  These bands are currently fixed a compile time.  It is intended for these ranges to be set by the end user at a later date.