binary of the first Alpha release of the TremorMouse
Just get yourself an RM2040 USB Host from adafruit.com  https://www.adafruit.com/product/5723

To load the binary hold down the Boot button and plug in the rp2040  to you computer.  The device should appear in your file manager.  Drag and drop the uf2 file to the device.  It will reset and disappear from you file manager.  Plug in you mouse.  Push and hold the Reset button for 1 second and you should be good to go.

At high speed the mouse movement is not attenuated.  At medium speed there is a 2;1 attenuation.  At slow speed the attenuation is 4:1

This will be configurable by the end user at a later date or if you are adventurous, compile you own configuration with the  Arduino IDE